requirements.txt
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.6.3
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
lxml==4.8.0
outcome==1.1.0
pycparser==2.21
pydantic==1.9.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.19.1
tornado==6.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.18.1
uvloop==0.16.0
watchfiles==0.15.0
webdriver-manager==3.6.1
webdrivermanager==0.10.0
websockets==10.3
wsproto==1.1.0
